feat(core): implemented mobile navigation at the bottom of the screen (#8)
* feat(core): created models for icon buttons and navigation links

Models allow us to provide the same schema in different components. A navigation link for example will be used in the main navigation bar and in the mobile bottom navigation bar. The models ensure, that these links will follow the same structure

* feat(core): implemented a safe html pipe

Rendering of SVGs and scripts will be blocked by angular by default. This pipe allows that the passed HTML (containing SVGs or scripts) won't be blocked

* feat(core): implemented a mobile navigation bar at the screen bottom

This navigation bar allows mobile users to navigate more comfortably. I had to apply a bottom margin in the app.component.html, because the bottom navbar is sticky and would otherwise overlap the main content

* refactor(core): adapted main navbar links to hide when mobile navbar is displayed

I adapted the css media breakpoints of the links in the main navigation bar to hide when the mobile navbar is being displayed

* chore: outsourced all core data to an own core module

* refactor(core): outsourced declaration of nav bar elements to own service

The mobile navigation icon buttons and the navigation links are now declared in navigation-bar-content.service.ts, where the data can be distributed to multiple components simultaneously. When the data in the service is being changed, the changes will update all components using this service at once