feat: imlpemented user registration form (#27)
* build: included the forms plugin for tailwindcss

* feat: created a module for administration

* chore: renamed a few components and models

* fix: changed hex color code of icons in the tailwind.config.js file

* style: standardized a few css properties of a and h3 elements

* feat: implemented an admin dashboard

This dashboard allows the prospective administrators to navigate between the admin views more comfortably. It contains a menu list consisting of links to the user registration, user management and event management

* style: replaced placeholder logo with match logo draft

* feat: implemented user registration page

* feat: created a shared button component

This component can be reused in different components, in order to simplify the development process. It can be included with the <app-shared-button> tag

* fix: fixed cyclic dependency look

The shared components were declared in the app root module, which was also declared in the admin module. This resulted in an infinite import root. Another error occurred, because the safe html pipe was also declared in multiple modules

* style: redesigned shared-button.component and made its design adaptive

I have made the button design adaptive. The style can be changed by passing either 'primary' or 'secondary' to the buttonDesign property

* feat: implemented html registration form

It is solely a plain html form. Its functionality still needs to be implemented in the type script file. The input fields should also be extracted to an own component

* feat: implemented reusable input component

* feat: implemented dynamic error messages of input component

The shared input component is now capable of showing an error message beneath the input field if a validator is not satisfied

* style: applied css styles when an input field is required, invalid or disabled

* feat: implemented reusable dropdown component

* refactor: replaced html form for user creation with a reactive form

* style: darkened the red if an input field is invalid

* style: adapted registration form grid to width of tablets and smartphones

* fix: input forms overlapped the nav bar when scrolling up

* style: introduced placeholder for the input fields of the registration form

* style: changed error messages from span to p element

Applying the left margin to a span element would only apply the margin to the span in the first line if it is a multiline span. Applying it to a p element includes all lines if it is a multiline element

* feat: implemented a validator for checking the format of the room number

* feat: implemented http client method calling the user creation endpoint

* feat: registration component now send its data to the backend

I have adapted the user registration component to send the form data to the backend over the rest api. I have also added a dropdown for selecting the gender, which is required for user registration

* style: improved arrangement of form input on tablet screen

* fix: label wasn't colored when field is invalid

I fixed this issue by declaring a custom tailwind utility class which is applied when the input or dropdown is invalid

* fix: deleting room number results in empty string (#21)

* fix: sets input form value to null again if typed in input was removed

* refactor(api): the createUser method now directly checks if the passed argument is a user object

* refactor: registrations form will now only be reset if request was performed successfully

* refactor: renamed and improved readability of the method for setting the default username

* feat: implemented notifications (#25)

* refactor: introduced own type for select options and marked final fields as readonly

* deps: installed ngx-toastr module for displaying notification messages

* feat: implemented notification service with custom notification component

The notification service wraps the toastr service and provides methods for displaying notification messages. I have also implemented a custom design component for the notification messages

* style: added duplicate counter to custom notification component and adapted design to desktop pcs

* style: added progress showing when notification disappears

* feat: added notification when request for registering users is being sent

* feat: implemented validation if username is already taken (#26)

* fix: enables register button only when from is valid

 Before that the button was enabled when the form state was pending

* feat: implemented a validator for checking if a username is available

* feat: implemented http call method for checking if a username is available

* feat: added form error messages for username availability checking

* fix: added required imports to component tests

* wip: tried to fix component test of the notification component

* fix(test): test required ToastPackage which couldn't be injected properly

In order to fix the issue, I have created an own ToastPackage object which will be used as ToastPackage provider